remove s3_servive_beta module
